rename investments pages for user and project, add btns to investments, add methods to investment controller, implemented delete functionality

+++ src/main/java/net/greatstart/controllers/InvestmentController.java
@@ -1,6 +1,7 @@
 package net.greatstart.controllers;
 
 import net.greatstart.dto.DtoInvestment;
+import net.greatstart.mappers.InvestmentMapper;
 import net.greatstart.model.Investment;
 import net.greatstart.model.Project;
 import net.greatstart.model.User;
@@ -9,13 +10,17 @@ import net.greatstart.services.ProjectService;
 import net.greatstart.services.UserService;
 import net.greatstart.validators.InvestmentValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -23,6 +28,7 @@ import java.math.BigDecimal;
 import java.security.Principal;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -30,12 +36,12 @@ import java.util.List;
 public class InvestmentController {
     private static final String PROJECT_PAGE = "project/";
     private static final String INVESTMENTS_VIEW = "investment/investments";
-    private static final String INVESTMENTS_PAGE = "investment";
     private static final String REDIRECT = "redirect:/";
     private static final String INVESTMENT_LIST = "investmentList";
     private static final String PAGE_NAME = "pageName";
 
     private InvestmentService investmentService;
+    private InvestmentMapper investmentMapper;
     private ProjectService projectService;
     private UserService userService;
     private InvestmentValidationService investmentValidationService;
@@ -80,6 +86,16 @@ public class InvestmentController {
         return model;
     }
 
+    @PutMapping("/api/investment/{id}")
+    public ResponseEntity<DtoInvestment> updateInvestment(@PathVariable long id,
+                                                          @RequestBody DtoInvestment investment) {
+        if (investmentService.getInvestmentById(id) != null) {
+            investmentService.deleteInvestment(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
     @GetMapping("/api/investment/{id}")
     @ResponseBody
     public DtoInvestment getInvestmentById(@PathVariable long id) {
@@ -95,22 +111,22 @@ public class InvestmentController {
         return model;
     }
 
-    @GetMapping("/api/investment/json")
+    @GetMapping("/api/investment")
     @ResponseBody
     public List<DtoInvestment> getInvestments() {
         return investmentService.getAllDtoInvestments();
     }
 
-    @GetMapping("/project/{id}/investments")
-    public ModelAndView getAllProjectInvestments(@PathVariable long id) {
-        ModelAndView model = new ModelAndView(INVESTMENTS_VIEW);
-
-        model.addObject(PAGE_NAME,
-                "Investments in project: "
-                        + projectService.getProjectById(id).getDesc().getName());
-        model.addObject(INVESTMENT_LIST,
-                projectService.getProjectById(id).getInvestments());
-        return model;
+    @GetMapping("/api/project/{id}/investments")
+    @ResponseBody
+    public ResponseEntity<List<DtoInvestment>> getAllProjectInvestments(@PathVariable long id) {
+        List<DtoInvestment> investments = new ArrayList<>();
+        projectService.getProjectById(id).getInvestments()
+                .forEach(investment -> investments.add(investmentMapper.fromInvestmentToDto(investment)));
+        if (investments.isEmpty()) {
+            return new ResponseEntity<>(investments, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/user/investments")
@@ -125,10 +141,14 @@ public class InvestmentController {
     }
 
 
-    @GetMapping("/investment/{id}/delete")
-    public ModelAndView deleteInvestmentById(@PathVariable long id) {
-        investmentService.deleteInvestment(id);
-        return new ModelAndView(REDIRECT + INVESTMENTS_PAGE);
+    @DeleteMapping("/api/investment/{id}")
+    @ResponseBody
+    public ResponseEntity<Investment> deleteInvestmentById(@PathVariable long id) {
+        if (investmentService.getInvestmentById(id) != null) {
+            investmentService.deleteInvestment(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     private BigDecimal getInvestedSumInProject(long id) {
